backend/Working server/requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.1
django-allauth==0.38.0
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
Markdown==3.0.1
oauthlib==2.1.0
psycopg2==2.7.5
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24
wincertstore==0.2
